% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getBattingData.R
\name{getPlayerSeasonData}
\alias{getPlayerSeasonData}
\title{Get player stats data from NPB official page}
\usage{

  getPlayerSeasonData(profile_page_url = "http://bis.npb.or.jp/players/51953886.html")
}
\arguments{
\item{batter}{profile page URL of NPB}
}
\value{
player data(data.frame)
}
\description{
Get player stats data from NPB official page
}

