% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/npbOfficialData.R
\docType{package}
\name{npbOfficialData}
\alias{npbOfficialData}
\alias{npbOfficialData-package}
\title{npbOfficialData
Get baseball player data from NPB}
\description{
npbOfficialData
Get baseball player data from NPB
}

